requirements.txt
openai
streamlit
pandas
matplotlib
python-dotenv
plotly
